+++ Il2CppInspector.Common/CppUtils/UnityHeaders/19-5.3.2.h
