+++ Il2CppInspector.Common/CppUtils/UnityHeaders/21-5.3.7-5.3.8.h
