+++ Il2CppInspector.Common/CppUtils/UnityHeaders/22-5.5.1-5.5.2.h
