+++ Il2CppInspector.Common/CppUtils/UnityHeaders/24-2017.1.3-2017.1.5.h
