+++ Il2CppInspector.Common/CppUtils/UnityHeaders/24.1-2018.3.8-2018.4.20.h
