+++ Il2CppInspector.GUI/MainWindow.xaml.cs
@@ -28,7 +28,7 @@ using Il2CppInspector.Outputs;
 using Il2CppInspector.Reflection;
 using Ookii.Dialogs.Wpf;
 using Path = System.IO.Path;
-using Il2CppInspector.Outputs.UnityHeaders;
+using Il2CppInspector.CppUtils.UnityHeaders;
 
 namespace Il2CppInspectorGUI
 {
