+++ Il2CppTests/TestIDAOutput.cs
@@ -6,7 +6,7 @@
 */
 
 using System.IO;
-using Il2CppInspector.Outputs.UnityHeaders;
+using Il2CppInspector.CppUtils.UnityHeaders;
 using Il2CppInspector.Reflection;
 using NUnit.Framework;
